Split builtins.py to several python files and try to follow Numpy's categories.
